Fixed boot API
By using @afourney's iOS boot request URL path, I was able to debug and
fix the /boot/ API within the golang implementation. BootHandler now
successfully makes a request to Pebble servers to retrieve the initial
bootstrap json file.

Comparing the output of the two systems, it was noticed that the GoLang
implementation was stripping out some of the keys out of the original
json file, unintentionally. This should have been corrected at this
point in time. We may end up missing things in the future, but if the
API never updates from this point forward, we should be fine.

The output of /boot/* has been modified to dump the output in a
tab-indented json format.