Added development info to README.md
Due to the helpful insight from @HenrydelMal, a backend development
section has been added to the README.md file. This should greatly help
those who are just starting to learn how to develop code in Go. As time
progresses, the document will eventually be merged with the copy that
resides in master.